app/requirements.txt
Flask==1.0.2
SQLAlchemy==1.3.3
SQLAlchemy_Utils==0.32.14
requests==2.18.4
python-dotenv==0.10.1
PyJWT==1.4.2
Flask-Cors==3.0.2
